Move lava files into lava-coder directory.
