+++ src/components/lava-coder/parse-ascii-raster.ts
